docs: Split community projects into sections, and add Java client (#758)
## Motivation and Context

I found the current raw assorted list of community project a bit
overwhelming and confusing at first.

So I gave some thought to it, and realized that at first I primarily
wanted to see what there is to browse the registry - so that should
probably go first? (The current magnifying glass icon, which was even
missing on one of the links, wasn't obvious enough, for me.)

Also I was looking for a Java client, and realized there already Go and
TS clients in this list - let's group those.

The Other section in the middle is for everything else (for now, it can
always further evolve, later).

## How Has This Been Tested?

N/A, as documentation only.

## Breaking Changes

Nope, it's documentation only.

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [X] Documentation update

## Checklist

- [X] I have read the [MCP
Documentation](https://modelcontextprotocol.io)
- [X] My code follows the repository's style guidelines
- [X] New and existing tests pass locally
- [X] I have added appropriate error handling
- [X] I have added or updated documentation as needed